Fixed two bugs.
Firstly, pstats.Stats(self.profiler) is not threadsafe when called with the
same profiler object from two threads. So we have wrapped it
in BProfile._reporting_lock to ensure only one runs at a time.

Futhermore, sometimes at exit time BProfile would call do_report from both
the atexit method, as well as the normal reporting loop. This sometimes
hit on a race condition in pstats.Stats, which as I mentioned is not threadsafe.
So now we wrap both of those methods' access to BProfile._instances_requiring_reports
in BProfile._reporting_lock as well, to ensure that their changes to that set
are atomic, and that only one of them will be able to call do_report on any
given profiler.

Small code cleanups.

Bumped patch version number